fix(#1909): PR #2016 — clear structural ratchet by extracting new hot-path helpers to siblings
Both files crossed their grandfathered #1995 ceilings:
- storage.ts grew 7061 -> 7302 (this PR's ContentHashIndex reconcile/rebuild
  locking + deferred durable-retry machinery)
- extraction-persist.ts grew 2998 -> 3013 (also pushed it over the 3000 oversize
  threshold, tripping oversizedFileCount 11 -> 12)

Move the additions into sibling modules — no behavior change, no baseline raise:

1. storage.ts: relocate the whole ContentHashIndex class plus its companions
   (ContentHashIndexLockOptions, FactHashIndexNotAuthoritativeError, the
   CONTENT_HASH_INDEX_* lock/retry constants) into a new
   storage/content-hash-index.ts (534 lines, under the 1200 new-file cap).
   storage.ts imports them back and re-exports the class + companion types so
   the public API stays stable for external importers (tests, dist). The
   FACT_HASH_INDEX_REBUILD_* constants and StorageManager methods that consume
   the index stay put; CONTENT_HASH_INDEX_RETRY_MAX_DELAY_MS is now exported and
   imported back for the one StorageManager rebuild-retry use. 7302 -> 6792.

2. extraction-persist.ts: move the new free helper readActiveMemoriesBothTiers
   into the existing extraction-persist-promotion.ts sibling (already the home
   of the extracted promotion/dedup gating helpers) and import it back.
   3013 -> 2993 (back under both the ceiling and the 3000 oversize threshold).

check-ratchets now passes (OK); baseline untouched. Core typecheck clean;
focused storage + extraction tests green, including the #2016 chunked-hash
durability regression.
